docs: create JSDocs for boards/*
